- Separate animation logic from protocol logic - Change date format to ISO 8601
